move Number::from_parts to Numeric
