+++ core/src/loader/gnu_units.rs
@@ -211,7 +211,7 @@ fn parse_term(iter: &mut Iter<'_>) -> Expr {
             }
             _ => Expr::new_unit(name),
         },
-        Token::Number(num, frac, exp) => crate::types::Number::from_parts(
+        Token::Number(num, frac, exp) => Numeric::from_parts(
             &*num,
             frac.as_ref().map(|x| &**x),
             exp.as_ref().map(|x| &**x),
